requirements.txt
Flask
python-dotenv
praw
transformers
torch
